initially setup state pattern
