+++ src/main/java/org/braekpo1nt/mctmanager/games/game/mecha/states/MechaState.java
@@ -1,0 +1,5 @@
+package org.braekpo1nt.mctmanager.games.game.mecha.states;
+
+public interface MechaState {
+    
+}
