+++ src/main/java/org/braekpo1nt/mctmanager/games/game/mecha/states/StartingState.java
@@ -1,0 +1,10 @@
+package org.braekpo1nt.mctmanager.games.game.mecha.states;
+
+import lombok.AllArgsConstructor;
+import org.braekpo1nt.mctmanager.games.game.mecha.MechaGame;
+
+@AllArgsConstructor
+public class StartingState implements MechaState {
+    private MechaGame context;
+    
+}
